requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1

